chore: Refactor Friday! Refactor the me reducer (#401)
* chore: moving from shared to me

* chore: add test for me reducer

* chore: use produce

* chore: run prettier one more time

* chore: add space that somehow wasn't picked up by prettier

* fix: change test per bucky's comment
