+++ src/me/actions/creators/index.ts
@@ -1,4 +1,4 @@
-import {MeState} from 'src/shared/reducers/me'
+import {MeState} from 'src/me/reducers'
 
 export const SET_ME = 'SET_ME'
 
@@ -7,7 +7,5 @@ export type Actions = ReturnType<typeof setMe>
 export const setMe = (me: MeState) =>
   ({
     type: SET_ME,
-    payload: {
-      me,
-    },
+    me,
   } as const)
