+++ src/store/configureStore.ts
@@ -1,6 +1,11 @@
-import {createStore, applyMiddleware, compose} from 'redux'
+import {
+  createStore,
+  applyMiddleware,
+  compose,
+  combineReducers,
+  Store,
+} from 'redux'
 import {History} from 'history'
-import {combineReducers, Store} from 'redux'
 import {connectRouter, routerMiddleware} from 'connected-react-router'
 import thunkMiddleware from 'redux-thunk'
 
@@ -12,7 +17,7 @@ import persistStateEnhancer from './persistStateEnhancer'
 import {loadLocalStorage} from 'src/localStorage'
 
 // v2 reducers
-import meReducer from 'src/shared/reducers/me'
+import meReducer from 'src/me/reducers'
 import flagReducer from 'src/shared/reducers/flags'
 import currentDashboardReducer from 'src/shared/reducers/currentDashboard'
 import currentExplorerReducer from 'src/shared/reducers/currentExplorer'
