+++ src/types/me.ts
@@ -1,0 +1,1 @@
+export {MeState} from 'src/me/reducers'
